Moving document to folder

+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/note_menu/data/model/Folder.kt
@@ -14,7 +14,7 @@ data class Folder(
     override val lastUpdatedAt: Instant,
     override val userId: String,
     override val favorite: Boolean = false,
-    val itemCount: Int,
+    val itemCount: Long,
     val documentList: List<Document> = emptyList(),
 ) : MenuItem {
     companion object {
