+++ application/features/note_menu/src/commonMain/kotlin/io/writeopia/note_menu/ui/dto/DocumentUi.kt
@@ -23,7 +23,7 @@ sealed interface MenuItemUi {
         override val title: String,
         override val selected: Boolean,
         override val isFavorite: Boolean,
-        val itemsCount: Int,
+        val itemsCount: Long,
         val parentId: String
     ) : MenuItemUi
 }
